+++ packages/Collapsible/src/Collapsible.styles.js
@@ -20,23 +20,24 @@ export const CollapsibleLabel = styled(RawButton)(
   ({ iconAlign, isToggleIconOnly }) => css`
     border-radius: ${tokens.border.radius};
     display: inline-block;
-    margin-left: -${tokens.spaceSm};
-    margin-right: -${tokens.spaceSm};
-    padding-left: ${tokens.spaceSm};
-    padding-right: ${tokens.spaceSm};
+    margin-inline-end: -${tokens.spaceSm};
+    margin-inline-start: -${tokens.spaceSm};
+    padding-inline-end: ${tokens.spaceSm};
+    padding-inline-start: ${tokens.spaceSm};
     width: calc(100% + ${tokens.space});
 
     ${isToggleIconOnly && `margin: 0; width: auto;`}
 
-    ${iconAlign === "left" && `margin-right: ${tokens.spaceSm};`}
-    ${iconAlign === "right" && `float: right; margin-left: ${tokens.spaceSm};`}
+    ${iconAlign === "left" && `margin-inline-end: ${tokens.spaceSm};`}
+    ${iconAlign === "right" && `float: inline-end; margin-inline-start: ${tokens.spaceSm};`}
   `
 );
 
 export const CollapsibleIcon = styled.span(
   ({ iconAlign }) => css`
-    ${iconAlign === "left" && `margin-right: ${tokens.spaceSm};`}
-    ${iconAlign === "right" && `float: right; margin-left: ${tokens.spaceSm};`}
+    ${iconAlign === "left" && `margin-inline-end: ${tokens.spaceSm};`}
+    ${iconAlign === "right" &&
+      `float: inline-end; margin-inline-start: ${tokens.spaceSm};`}
 
     [data-pka-anchor="icon"] {
       color: ${tokens.textColor.icon};