+++ packages/DatePicker/test/DatePicker.cypress.js
@@ -51,10 +51,10 @@ describe("<DatePicker />", () => {
     cy.findByTestId("calendar.shortcut").should("be.visible");
 
     cy.get('label[for$="-0"]').click();
-    cy.get('label[for$="-2018"]').click();
+    cy.get('label[for$="-2024"]').click();
     cy.findByTestId("calendar.apply").click();
     cy.findAllByTestId("calendar.header")
-      .contains("January 2018")
+      .contains("January 2024")
       .should("be.visible");
   });