+++ packages/NotificationCard/src/components/Image/Image.styles.js
@@ -14,7 +14,7 @@ export const NotificationImage = styled.div`
   flex-basis: 60%;
   max-width: ${BASE_WIDTH + stylers.toInt(tokens.space) * 3}px;
   min-width: ${BASE_WIDTH / 2}px;
-  padding-right: ${stylers.spacer(3)};
+  padding-inline-end: ${stylers.spacer(3)};
   width: 100%;
 
   ${({ align }) => (align === "center" ? alignedCenterStyles : "")}