+++ packages/Pagination/src/components/Wrapper/Wrapper.styles.js
@@ -2,9 +2,9 @@ import styled from "styled-components";
 import tokens from "@paprika/tokens";
 
 const Wrapper = styled.span`
-  border-right: 1px solid ${tokens.border.color};
-  margin-right: ${tokens.spaceSm};
-  padding-right: ${tokens.spaceSm};
+  border-inline-end: 1px solid ${tokens.border.color};
+  margin-inline-end: ${tokens.spaceSm};
+  padding-inline-end: ${tokens.spaceSm};
   position: relative;
 `;