+++ packages/Card/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 2.0.9
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0
+  - @paprika/heading@1.0.19
+  - @paprika/stylers@1.1.11
+
+## 2.0.9-next.0
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0-next.0
+  - @paprika/heading@1.0.19-next.0
+  - @paprika/stylers@1.1.11-next.0
+
 ## 2.0.8
 
 ### Patch Changes