+++ packages/CollapsibleChecklists/CHANGELOG.md
@@ -1,5 +1,49 @@
 # Changelog
 
+## 1.0.37
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0
+  - @paprika/checkbox@1.0.28
+  - @paprika/collapsible@4.0.2
+  - @paprika/stylers@1.1.11
+
+## 1.0.37-next.0
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0-next.0
+  - @paprika/checkbox@1.0.28-next.0
+  - @paprika/collapsible@4.0.2-next.0
+  - @paprika/stylers@1.1.11-next.0
+
+## 1.0.36
+
+### Patch Changes
+
+- @paprika/collapsible@4.0.1
+
+## 1.0.36-next.0
+
+### Patch Changes
+
+- @paprika/collapsible@4.0.1-next.0
+
+## 1.0.35
+
+### Patch Changes
+
+- @paprika/collapsible@4.0.0
+
+## 1.0.35-next.0
+
+### Patch Changes
+
+- @paprika/collapsible@4.0.0-next.0
+
 ## 1.0.34
 
 ### Patch Changes