+++ packages/DataField/CHANGELOG.md
@@ -1,5 +1,59 @@
 <!-- content -->
 
+## 4.0.2
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0
+  - @paprika/button@1.1.19
+  - @paprika/data-grid@4.0.2
+  - @paprika/icon@3.0.2
+  - @paprika/popover@2.0.11
+
+## 4.0.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0-next.0
+  - @paprika/button@1.1.19-next.0
+  - @paprika/data-grid@4.0.2-next.0
+  - @paprika/icon@3.0.2-next.0
+  - @paprika/popover@2.0.11-next.0
+
+## 4.0.1
+
+### Patch Changes
+
+- Updated dependencies [48a369d]
+  - @paprika/l10n@2.1.1
+  - @paprika/data-grid@4.0.1
+
+## 4.0.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [48a369d]
+  - @paprika/l10n@2.1.1-next.0
+  - @paprika/data-grid@4.0.1-next.0
+
+## 4.0.0
+
+### Patch Changes
+
+- Updated dependencies [1ec4f6b]
+  - @paprika/l10n@2.1.0
+  - @paprika/data-grid@4.0.0
+
+## 4.0.0-next.0
+
+### Patch Changes
+
+- Updated dependencies [1ec4f6b]
+  - @paprika/l10n@2.1.0-next.0
+  - @paprika/data-grid@4.0.0-next.0
+
 ## 3.0.3
 
 ### Patch Changes