+++ packages/DynamicHyperlinkTransformer/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @paprika/dynamic-hyperlink-transformer
 
+## 4.0.2
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0
+  - @paprika/stylers@1.1.11
+  - @paprika/tabs@2.3.12
+
+## 4.0.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0-next.0
+  - @paprika/stylers@1.1.11-next.0
+  - @paprika/tabs@2.3.12-next.0
+
+## 4.0.1
+
+### Patch Changes
+
+- Updated dependencies [48a369d]
+  - @paprika/l10n@2.1.1
+
+## 4.0.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [48a369d]
+  - @paprika/l10n@2.1.1-next.0
+
+## 4.0.0
+
+### Patch Changes
+
+- Updated dependencies [1ec4f6b]
+  - @paprika/l10n@2.1.0
+
+## 4.0.0-next.0
+
+### Patch Changes
+
+- Updated dependencies [1ec4f6b]
+  - @paprika/l10n@2.1.0-next.0
+
 ## 3.0.1
 
 ### Patch Changes