+++ packages/InlineEditors/CHANGELOG.md
@@ -1,5 +1,81 @@
 # @paprika/inline-editors
 
+## 1.0.21
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0
+  - @paprika/collapsible-text@4.0.2
+  - @paprika/icon@3.0.2
+  - @paprika/input@7.0.2
+  - @paprika/list-box@7.0.2
+  - @paprika/popover@2.0.11
+  - @paprika/raw-button@1.0.18
+  - @paprika/spinner@1.0.22
+  - @paprika/stylers@1.1.11
+  - @paprika/table@1.2.34
+  - @paprika/tag@3.0.2
+  - @paprika/textarea@3.0.14
+
+## 1.0.21-next.0
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0-next.0
+  - @paprika/collapsible-text@4.0.2-next.0
+  - @paprika/icon@3.0.2-next.0
+  - @paprika/input@7.0.2-next.0
+  - @paprika/list-box@7.0.2-next.0
+  - @paprika/popover@2.0.11-next.0
+  - @paprika/raw-button@1.0.18-next.0
+  - @paprika/spinner@1.0.22-next.0
+  - @paprika/stylers@1.1.11-next.0
+  - @paprika/table@1.2.34-next.0
+  - @paprika/tag@3.0.2-next.0
+  - @paprika/textarea@3.0.14-next.0
+
+## 1.0.20
+
+### Patch Changes
+
+- @paprika/collapsible-text@4.0.1
+- @paprika/input@7.0.1
+- @paprika/list-box@7.0.1
+- @paprika/table@1.2.33
+- @paprika/tag@3.0.1
+
+## 1.0.20-next.0
+
+### Patch Changes
+
+- @paprika/collapsible-text@4.0.1-next.0
+- @paprika/input@7.0.1-next.0
+- @paprika/list-box@7.0.1-next.0
+- @paprika/tag@3.0.1-next.0
+- @paprika/table@1.2.33-next.0
+
+## 1.0.19
+
+### Patch Changes
+
+- @paprika/collapsible-text@4.0.0
+- @paprika/input@7.0.0
+- @paprika/list-box@7.0.0
+- @paprika/tag@3.0.0
+- @paprika/table@1.2.32
+
+## 1.0.19-next.0
+
+### Patch Changes
+
+- @paprika/collapsible-text@4.0.0-next.0
+- @paprika/input@7.0.0-next.0
+- @paprika/list-box@7.0.0-next.0
+- @paprika/tag@3.0.0-next.0
+- @paprika/table@1.2.32-next.0
+
 ## 1.0.18
 
 ### Patch Changes