+++ packages/Popover/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 2.0.11
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0
+  - @paprika/raw-button@1.0.18
+  - @paprika/stylers@1.1.11
+
+## 2.0.11-next.0
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0-next.0
+  - @paprika/raw-button@1.0.18-next.0
+  - @paprika/stylers@1.1.11-next.0
+
 ## 2.0.10
 
 ### Patch Changes