+++ packages/Radio/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 1.0.30
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0
+  - @paprika/icon@3.0.2
+  - @paprika/stylers@1.1.11
+
+## 1.0.30-next.0
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0-next.0
+  - @paprika/icon@3.0.2-next.0
+  - @paprika/stylers@1.1.11-next.0
+
 ## 1.0.29
 
 ### Patch Changes