+++ packages/StatusTracker/CHANGELOG.md
@@ -1,5 +1,61 @@
 # Changelog
 
+## 5.0.2
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0
+  - @paprika/button@1.1.19
+  - @paprika/overflow-menu@5.0.2
+  - @paprika/popover@2.0.11
+  - @paprika/raw-button@1.0.18
+  - @paprika/stylers@1.1.11
+
+## 5.0.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0-next.0
+  - @paprika/button@1.1.19-next.0
+  - @paprika/overflow-menu@5.0.2-next.0
+  - @paprika/popover@2.0.11-next.0
+  - @paprika/raw-button@1.0.18-next.0
+  - @paprika/stylers@1.1.11-next.0
+
+## 5.0.1
+
+### Patch Changes
+
+- Updated dependencies [48a369d]
+  - @paprika/l10n@2.1.1
+  - @paprika/overflow-menu@5.0.1
+
+## 5.0.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [48a369d]
+  - @paprika/l10n@2.1.1-next.0
+  - @paprika/overflow-menu@5.0.1-next.0
+
+## 5.0.0
+
+### Patch Changes
+
+- Updated dependencies [1ec4f6b]
+  - @paprika/l10n@2.1.0
+  - @paprika/overflow-menu@5.0.0
+
+## 5.0.0-next.0
+
+### Patch Changes
+
+- Updated dependencies [1ec4f6b]
+  - @paprika/l10n@2.1.0-next.0
+  - @paprika/overflow-menu@5.0.0-next.0
+
 ## 4.0.1
 
 ### Patch Changes