+++ packages/Stylers/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @paprika/stylers
 
+## 1.1.11
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0
+
+## 1.1.11-next.0
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0-next.0
+
 ## 1.1.10
 
 ### Patch Changes