+++ packages/Textarea/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 3.0.14
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0
+  - @paprika/stylers@1.1.11
+
+## 3.0.14-next.0
+
+### Patch Changes
+
+- Updated dependencies [80fb5c0]
+  - @paprika/tokens@3.1.0-next.0
+  - @paprika/stylers@1.1.11-next.0
+
 ## 3.0.13
 
 ### Patch Changes